backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.12.1
ruff==0.1.8
structlog==23.2.0
python-magic==0.4.27
redis==5.0.1
rq==1.15.1